@@ -1,32 +1,45 @@
 using Authorization.Sql;
-using Autofac.Extensions.DependencyInjection;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
+using Serilog;
+using Serilog.Events;
+using Serilog.Exceptions;
+using Serilog.Exceptions.Core;
 
-namespace Authorization
+namespace Authorization;
+
+public class Program
 {
-    public class Program
+    public static void Main(string[] args)
     {
-        public static void Main(string[] args)
-        {
-            var host = CreateHostBuilder(args).Build();
-            MigrationTool.Execute(host.Services);
-            host.Run();
-        }
-
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureAppConfiguration((hostingContext, config) =>
-                {
-                    config.AddJsonFile("appsettings.json", false, true);
-                    config.AddEnvironmentVariables();
-                })
-                .UseServiceProviderFactory(new AutofacServiceProviderFactory())
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder.ConfigureServices(services => services.AddAutofac());
-                    webBuilder.UseStartup<Startup>();
-                });
+        var host = CreateHostBuilder(args).Build();
+        MigrationTool.Execute(host.Services);
+        host.Run();
     }
+
+    public static IHostBuilder CreateHostBuilder(string[] args) =>
+        Host.CreateDefaultBuilder(args)
+            .ConfigureAppConfiguration((hostingContext, config) =>
+            {
+                var env = hostingContext.HostingEnvironment;
+
+                config.AddJsonFile("appsettings.json", false, true)
+                    .AddJsonFile($"appsettings.{env.EnvironmentName}.json", true, true);
+
+                config.AddEnvironmentVariables();
+            })
+            .UseSerilog((ctx, _, cfg) =>
+                cfg
+                    .Enrich.FromLogContext()
+                    .Enrich.WithExceptionDetails(new DestructuringOptionsBuilder())
+                    .Enrich.WithThreadId()
+                    .Enrich.WithThreadName()
+                    .ReadFrom.Configuration(ctx.Configuration)
+                    .MinimumLevel.Override("System", LogEventLevel.Information)
+            )
+            .ConfigureWebHostDefaults(webBuilder =>
+            {
+                webBuilder.UseStartup<Startup>();
+            });
 }
